Simplifies locks
Reduce the number of locks, but tries to put related things in the same
lock. It avoid that an operation depends on another, but under a
different lock. When the structure changes, bad things could happen.

Lock only in public methods. As write locks are not reentrant, avoid
the structure to lock itself